libsc update
get sc_shmem_array feature so that it can be used in p4est for arrays
that are redundant on every process, particularly those that are O(P).
This can potentially reduce the memory devoted to these arrays by a
factor of S, where S is the number of mpi processes in a shared memory
region.

libsc update

libsc update

libsc update

libsc update

libsc update