p4est_balance_sort: start developing neighbor comm tree
We want to avoid too many sends/recvs on one process in ghost neighbor
communication.  A processes can have a number of neighbors dependent on
how refined the tree is (difference between maximum and minimum quadrant
size).  Coalescing messages can combat this.  When can we / should we
coalesce messages?

One observation: the way that a process has lots of neighbors is that it
has lots of smaller neighbors.  Assertion: the number of larger/equal
sized neighbors that a process can have is bounded (the bound depends on
the forest topology).  If the assertion is true, then we can bound the
number of sends/recvs on any one process by coalescing messages that go
from smaller processes to bigger processes.

The idea is to coalesce in a tree pattern.  I've introduced a parameter
for the tree branching degree.  It is not used yet, but if the tree
branching parameter is non-zero, then statistics are collected about how
many larger/same/smaller neighbors each process has.  In
p8est_test_fusion I do observe that, taking the same highly refined tree
and changing the number of processes, larger/same statistics stay
bounded, while smaller grows.